Move issue-submission CLI to MEDS_DEV.web; read body from env var directly
Three review notes on #295, all the same direction:

- mmcdermott: "Does this really belong in results, or should it go in
  the web submodule?" — yes, web. The extraction-and-write logic is part
  of the GitHub-issue submission flow, which is a web/submission concern.
  Result.from_dict stays in results (it's a general deserializer).
- mmcdermott: "I'm not sure if this name is right — this is more than
  just validation. It's parse, validate, and write." — agreed.
  Renamed: validate_result_from_issue -> process_submission. The CLI
  becomes meds-dev-process-submission.
- mmcdermott: "Why the printf pipe at all?" — fair, was just for env
  var routing. The new CLI reads $ISSUE_BODY directly so the workflow
  no longer needs a pipe.

Structurally:

- New module MEDS_DEV.web.process_submission with:
  - extract_result_dict_from_issue_body(body)  (moved from results)
  - process_submission(body, result_fp, do_overwrite)
  - main()  — CLI: --issue_body_fp > $ISSUE_BODY > stdin
- MEDS_DEV.results loses the extractor and the CLI; Result.from_dict
  stays put (general dict-deserializer).
- pyproject.toml: meds-dev-validate-result-from-issue is gone, replaced
  by meds-dev-process-submission. meds-dev-validate-result is preserved
  for ad-hoc file-in validation (e.g., local checks before submission).
- upload_benchmark_result.yaml: the printf pipe is gone. The CLI reads
  ISSUE_BODY from the env block directly.
