+++ .github/workflows/upload_benchmark_result.yaml
@@ -24,12 +24,12 @@ jobs:
 
       - name: Extract and validate result from issue body
         env:
+          # The CLI reads the body from $ISSUE_BODY by default. Routing through an env var
+          # avoids shell injection — putting ${{ github.event.issue.body }} directly into a
+          # command line would let untrusted issue content escape and execute.
           ISSUE_BODY: ${{ github.event.issue.body }}
         run: |
-          # Single Python invocation: parses the ```json``` block out of the issue body,
-          # constructs a Result (which validates via Result.__post_init__), and writes the
-          # canonical (NaN-sanitized) JSON to disk.
-          printf '%s' "$ISSUE_BODY" | meds-dev-validate-result-from-issue \
+          meds-dev-process-submission \
             --result_fp "$GITHUB_WORKSPACE/result.json" \
             --do_overwrite
