+++ pyproject.toml
@@ -65,9 +65,9 @@ meds-dev-model = "MEDS_DEV.models.__main__:main"
 meds-dev-evaluation = "MEDS_DEV.evaluation.__main__:main"
 meds-dev-pack-result = "MEDS_DEV.results.__main__:pack_result"
 meds-dev-validate-result = "MEDS_DEV.results.__main__:validate_result"
-meds-dev-validate-result-from-issue = "MEDS_DEV.results.__main__:validate_result_from_issue"
 meds-dev-collate-entities = "MEDS_DEV.web.collate_entities:main"
 meds-dev-aggregate-results = "MEDS_DEV.web.aggregate_results:main"
+meds-dev-process-submission = "MEDS_DEV.web.process_submission:main"
 
 [project.urls]
 Homepage = "https://github.com/Medical-Event-Data-Standard/MEDS-DEV"
