+++ src/MEDS_DEV/web/__init__.py
@@ -1,6 +1,12 @@
-"""Web tooling: collate entities and aggregate results for the MEDS-DEV website."""
+"""Web tooling: collate entities, aggregate results, and process result submissions."""
 
 from .aggregate_results import aggregate_results
 from .collate_entities import collate_entities
+from .process_submission import extract_result_dict_from_issue_body, process_submission
 
-__all__ = ["aggregate_results", "collate_entities"]
+__all__ = [
+    "aggregate_results",
+    "collate_entities",
+    "extract_result_dict_from_issue_body",
+    "process_submission",
+]
